- implement DROP TABLE support, toggled off from generation for now - clean up the query generation/printing by separating it into different files and removing duplications
