+++ simulator/generation/property.rs
@@ -3,7 +3,10 @@ use serde::{Deserialize, Serialize};
 
 use crate::{
     model::{
-        query::{Create, Delete, Distinctness, Insert, Predicate, Query, Select},
+        query::{
+            select::{Distinctness, Predicate},
+            Create, Delete, Drop, Insert, Query, Select,
+        },
         table::Value,
     },
     runner::env::SimulatorEnv,
@@ -94,6 +97,23 @@ pub(crate) enum Property {
         predicate: Predicate,
         queries: Vec<Query>,
     },
+    // Drop-Select is a property in which selecting from a dropped table
+    // should result in an error.
+    // The execution of the property is as follows
+    //     DROP TABLE <t>
+    //     I_0
+    //     I_1
+    //     ...
+    //     I_n
+    //     SELECT * FROM <t> WHERE <predicate> -> Error
+    // The interactions in the middle has the following constraints;
+    // - There will be no errors in the middle interactions.
+    // - The table `t` will not be created, no table will be renamed to `t`.
+    DropSelect {
+        table: String,
+        queries: Vec<Query>,
+        select: Select,
+    },
 }
 
 impl Property {
@@ -103,6 +123,7 @@ impl Property {
             Property::DoubleCreateFailure { .. } => "Double-Create-Failure".to_string(),
             Property::SelectLimit { .. } => "Select-Limit".to_string(),
             Property::DeleteSelect { .. } => "Delete-Select".to_string(),
+            Property::DropSelect { .. } => "Drop-Select".to_string(),
         }
     }
     /// interactions construct a list of interactions, which is an executable representation of the property.
@@ -289,6 +310,55 @@ impl Property {
 
                 interactions
             }
+            Property::DropSelect {
+                table,
+                queries,
+                select,
+            } => {
+                let assumption = Interaction::Assumption(Assertion {
+                    message: format!("table {} exists", table),
+                    func: Box::new({
+                        let table = table.clone();
+                        move |_: &Vec<ResultSet>, env: &SimulatorEnv| {
+                            Ok(env.tables.iter().any(|t| t.name == table))
+                        }
+                    }),
+                });
+
+                let table_name = table.clone();
+
+                let assertion = Interaction::Assertion(Assertion {
+                    message: format!(
+                        "select query should result in an error for table '{}'",
+                        table
+                    ),
+                    func: Box::new(move |stack: &Vec<ResultSet>, _: &SimulatorEnv| {
+                        let last = stack.last().unwrap();
+                        match last {
+                            Ok(_) => Ok(false),
+                            Err(e) => Ok(e
+                                .to_string()
+                                .contains(&format!("Table {table_name} does not exist"))),
+                        }
+                    }),
+                });
+
+                let drop = Interaction::Query(Query::Drop(Drop {
+                    table: table.clone(),
+                }));
+
+                let select = Interaction::Query(Query::Select(select.clone()));
+
+                let mut interactions = Vec::new();
+
+                interactions.push(assumption);
+                interactions.push(drop);
+                interactions.extend(queries.clone().into_iter().map(Interaction::Query));
+                interactions.push(select);
+                interactions.push(assertion);
+
+                interactions
+            }
         }
     }
 }
@@ -298,6 +368,8 @@ pub(crate) struct Remaining {
     pub(crate) read: f64,
     pub(crate) write: f64,
     pub(crate) create: f64,
+    pub(crate) delete: f64,
+    pub(crate) drop: f64,
 }
 
 pub(crate) fn remaining(env: &SimulatorEnv, stats: &InteractionStats) -> Remaining {
@@ -310,11 +382,19 @@ pub(crate) fn remaining(env: &SimulatorEnv, stats: &InteractionStats) -> Remaini
     let remaining_create = ((env.opts.max_interactions as f64 * env.opts.create_percent / 100.0)
         - (stats.create_count as f64))
         .max(0.0);
+    let remaining_delete = ((env.opts.max_interactions as f64 * env.opts.delete_percent / 100.0)
+        - (stats.delete_count as f64))
+        .max(0.0);
+    let remaining_drop = ((env.opts.max_interactions as f64 * env.opts.drop_percent / 100.0)
+        - (stats.drop_count as f64))
+        .max(0.0);
 
     Remaining {
         read: remaining_read,
         write: remaining_write,
         create: remaining_create,
+        delete: remaining_delete,
+        drop: remaining_drop,
     }
 }
 
@@ -479,6 +559,47 @@ fn property_delete_select<R: rand::Rng>(
         queries,
     }
 }
+
+fn property_drop_select<R: rand::Rng>(
+    rng: &mut R,
+    env: &SimulatorEnv,
+    remaining: &Remaining,
+) -> Property {
+    // Get a random table
+    let table = pick(&env.tables, rng);
+
+    // Create random queries respecting the constraints
+    let mut queries = Vec::new();
+    // - [x] There will be no errors in the middle interactions. (this constraint is impossible to check, so this is just best effort)
+    // - [-] The table `t` will not be created, no table will be renamed to `t`. (todo: update this constraint once ALTER is implemented)
+    for _ in 0..rng.gen_range(0..3) {
+        let query = Query::arbitrary_from(rng, (env, remaining));
+        match &query {
+            Query::Create(Create { table: t }) => {
+                // - The table `t` will not be created
+                if t.name == table.name {
+                    continue;
+                }
+            }
+            _ => (),
+        }
+        queries.push(query);
+    }
+
+    let select = Select {
+        table: table.name.clone(),
+        predicate: Predicate::arbitrary_from(rng, table),
+        limit: None,
+        distinct: Distinctness::All,
+    };
+
+    Property::DropSelect {
+        table: table.name.clone(),
+        queries,
+        select,
+    }
+}
+
 impl ArbitraryFrom<(&SimulatorEnv, &InteractionStats)> for Property {
     fn arbitrary_from<R: rand::Rng>(
         rng: &mut R,
@@ -502,6 +623,11 @@ impl ArbitraryFrom<(&SimulatorEnv, &InteractionStats)> for Property {
                 (
                     f64::min(remaining_.read, remaining_.write),
                     Box::new(|rng: &mut R| property_delete_select(rng, env, &remaining_)),
+                ),
+                (
+                    // remaining_.drop,
+                    0.0,
+                    Box::new(|rng: &mut R| property_drop_select(rng, env, &remaining_)),
                 ),
             ],
             rng,
