+++ simulator/generation/query.rs
@@ -1,7 +1,8 @@
 use crate::generation::table::{GTValue, LTValue};
 use crate::generation::{one_of, Arbitrary, ArbitraryFrom};
 
-use crate::model::query::{Create, Delete, Distinctness, Insert, Predicate, Query, Select};
+use crate::model::query::select::{Distinctness, Predicate};
+use crate::model::query::{Create, Delete, Drop, Insert, Query, Select};
 use crate::model::table::{Table, Value};
 use crate::SimulatorEnv;
 use rand::seq::SliceRandom as _;
@@ -92,6 +93,15 @@ impl ArbitraryFrom<&SimulatorEnv> for Delete {
         Self {
             table: table.name.clone(),
             predicate: Predicate::arbitrary_from(rng, table),
+        }
+    }
+}
+
+impl ArbitraryFrom<&SimulatorEnv> for Drop {
+    fn arbitrary_from<R: Rng>(rng: &mut R, env: &SimulatorEnv) -> Self {
+        let table = pick(&env.tables, rng);
+        Self {
+            table: table.name.clone(),
         }
     }
 }
